유저 이메일 관련 코드 정리 (#159)
- 뮤테이션 및 이메일 템플릿 이름 컨벤션 맞춤
- auth에서 user로 관련 뮤테이션 옮김
