+++ apps/api/src/email/templates/index.ts
@@ -1,4 +1,4 @@
-export { default as EmailChangeEmail } from './EmailChangeEmail';
-export { default as NotifyEmailChangeEmail } from './NotifyEmailChangeEmail';
+export { default as EmailUpdatedEmail } from './EmailUpdatedEmail';
+export { default as EmailUpdateEmail } from './EmailUpdateEmail';
 export { default as PasswordResetEmail } from './PasswordResetEmail';
 export { default as SignUpEmail } from './SignUpEmail';
