+++ apps/api/src/graphql/resolvers/user.ts
@@ -1,3 +1,13 @@
+import { and, eq } from 'drizzle-orm';
+import { nanoid } from 'nanoid';
+import { redis } from '@/cache';
+import { db, first, firstOrThrow } from '@/db';
+import { Users } from '@/db/schemas/tables';
+import { sendEmail } from '@/email';
+import { EmailUpdatedEmail, EmailUpdateEmail } from '@/email/templates';
+import { UserState } from '@/enums';
+import { env } from '@/env';
+import { TypieError } from '@/errors';
 import { builder } from '../builder';
 import { User } from '../objects';
 
@@ -23,6 +33,89 @@ builder.queryFields((t) => ({
     nullable: true,
     resolve: async (_, __, ctx) => {
       return ctx.session?.userId;
+    },
+  }),
+}));
+
+/**
+ * * Mutations
+ */
+
+builder.mutationFields((t) => ({
+  sendEmailUpdateEmail: t.withAuth({ session: true }).fieldWithInput({
+    type: 'Boolean',
+    input: { email: t.input.string() },
+    resolve: async (_, { input }, ctx) => {
+      const email = input.email.toLowerCase();
+
+      const existingUser = await db
+        .select({ id: Users.id })
+        .from(Users)
+        .where(and(eq(Users.email, email), eq(Users.state, UserState.ACTIVE)))
+        .then(first);
+
+      if (existingUser) {
+        throw new TypieError({ code: 'user_email_exists' });
+      }
+
+      const user = await db.select({ name: Users.name }).from(Users).where(eq(Users.id, ctx.session.userId)).then(firstOrThrow);
+
+      const code = nanoid();
+
+      await redis.setex(
+        `user:update-email:${code}`,
+        24 * 60 * 60,
+        JSON.stringify({
+          email,
+          userId: ctx.session.userId,
+        }),
+      );
+
+      await sendEmail({
+        recipient: input.email,
+        subject: '[타이피] 이메일 주소를 인증해 주세요',
+        body: EmailUpdateEmail({
+          name: user.name,
+          email,
+          verificationUrl: `${env.WEBSITE_URL}/auth/update-email?code=${code}`,
+        }),
+      });
+
+      return true;
+    },
+  }),
+
+  updateEmail: t.fieldWithInput({
+    type: 'Boolean',
+    input: { code: t.input.string() },
+    resolve: async (_, { input }) => {
+      const data = await redis.get(`user:update-email:${input.code}`);
+      if (!data) {
+        throw new TypieError({ code: 'invalid_code' });
+      }
+
+      const { email, userId } = JSON.parse(data);
+
+      const user = await db
+        .select({ id: Users.id, name: Users.name, email: Users.email })
+        .from(Users)
+        .where(and(eq(Users.id, userId), eq(Users.state, UserState.ACTIVE)))
+        .then(firstOrThrow);
+
+      await db.update(Users).set({ email }).where(eq(Users.id, user.id));
+
+      await redis.del(`user:update-email:${input.code}`);
+
+      await sendEmail({
+        recipient: user.email,
+        subject: '[타이피] 이메일 주소가 변경되었어요',
+        body: EmailUpdatedEmail({
+          name: user.name,
+          email,
+        }),
+      });
+
+      return true;
     },
   }),
 }));
